NP-1387: move exceptions to own package
