+++ user-access-errors/src/test/java/no/unit/nva/useraccessmanagement/exceptions/BadRequestExceptionTest.java
@@ -1,8 +1,9 @@
-package no.unit.nva.exceptions;
+package no.unit.nva.useraccessmanagement.exceptions;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
+import no.unit.nva.useraccessmanagement.exceptions.BadRequestException;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
