+++ user-access-errors/src/test/java/no/unit/nva/useraccessmanagement/exceptions/EmptyInputExceptionTest.java
@@ -1,9 +1,10 @@
-package no.unit.nva.exceptions;
+package no.unit.nva.useraccessmanagement.exceptions;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
+import no.unit.nva.useraccessmanagement.exceptions.EmptyInputException;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
