+++ user-access-errors/src/test/java/no/unit/nva/useraccessmanagement/exceptions/InvalidInputExceptionTest.java
@@ -1,10 +1,10 @@
-package no.unit.nva.exceptions;
+package no.unit.nva.useraccessmanagement.exceptions;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
-import no.unit.nva.useraccessmanagement.model.exceptions.InvalidInputException;
+import no.unit.nva.useraccessmanagement.exceptions.InvalidInputException;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.Test;
