+++ user-access-handlers/src/test/java/no/unit/nva/handlers/AddRoleHandlerTest.java
@@ -15,7 +15,7 @@ import java.io.InputStream;
 import no.unit.nva.database.DatabaseService;
 import no.unit.nva.database.DatabaseServiceImpl;
 import no.unit.nva.database.RoleService;
-import no.unit.nva.exceptions.DataSyncException;
+import no.unit.nva.useraccessmanagement.exceptions.DataSyncException;
 import no.unit.nva.useraccessmanagement.exceptions.InvalidEntryInternalException;
 import no.unit.nva.testutils.HandlerRequestBuilder;
 import no.unit.nva.useraccessmanagement.model.RoleDto;
