+++ user-access-handlers/src/test/java/no/unit/nva/handlers/AddUserTest.java
@@ -25,10 +25,10 @@ import java.lang.reflect.InvocationTargetException;
 import no.unit.nva.database.DatabaseService;
 import no.unit.nva.database.DatabaseServiceImpl;
 
-import no.unit.nva.exceptions.DataSyncException;
+import no.unit.nva.useraccessmanagement.exceptions.DataSyncException;
 import no.unit.nva.useraccessmanagement.exceptions.InvalidEntryInternalException;
 import no.unit.nva.useraccessmanagement.model.UserDto;
-import no.unit.nva.useraccessmanagement.model.exceptions.InvalidInputException;
+import no.unit.nva.useraccessmanagement.exceptions.InvalidInputException;
 
 import nva.commons.exceptions.ApiGatewayException;
 import nva.commons.exceptions.InvalidOrMissingTypeException;
