+++ user-access-handlers/src/test/java/no/unit/nva/handlers/UpdateUserHandlerTest.java
@@ -25,7 +25,7 @@ import no.unit.nva.testutils.HandlerRequestBuilder;
 import no.unit.nva.useraccessmanagement.exceptions.InvalidEntryInternalException;
 import no.unit.nva.useraccessmanagement.model.RoleDto;
 import no.unit.nva.useraccessmanagement.model.UserDto;
-import no.unit.nva.useraccessmanagement.model.exceptions.InvalidInputException;
+import no.unit.nva.useraccessmanagement.exceptions.InvalidInputException;
 import nva.commons.exceptions.ApiGatewayException;
 import nva.commons.exceptions.InvalidOrMissingTypeException;
 import nva.commons.exceptions.commonexceptions.ConflictException;
