+++ user-access-public-model/src/test/java/no/unit/nva/useraccessmanagement/model/EntityUtils.java
@@ -120,6 +120,13 @@ public final class EntityUtils {
             .build();
     }
 
+    /**
+     * Creates a sample role.
+     *
+     * @param someRole role name.
+     * @return a sample role.
+     * @throws InvalidEntryInternalException when the generated role is invalid.
+     */
     public static RoleDto createRole(String someRole) throws InvalidEntryInternalException {
         return
             RoleDto.newBuilder()
